refactor(database): decompose the users god-file by role (WS-D)
Split internal/database/users.go (731 LOC) into role-based files in the
same package (following the existing users_*.go convention) — pure move:
core user CRUD stays in users.go; login/lockout/token-version moves to
users_auth.go; SSO upsert/lookup moves to users_sso.go.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
